requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt --strip-extras requirements.in
#
attrs==25.3.0
    # via
    #   flake8-bugbear
    #   outcome
    #   trio
black==24.10.0
    # via
    #   -r requirements.in
    #   flake8-black
certifi==2025.1.31
    # via
    #   requests
    #   selenium
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via black
distlib==0.3.9
    # via virtualenv
filelock==3.18.0
    # via virtualenv
flake8==7.1.1
    # via
    #   -r requirements.in
    #   flake8-black
    #   flake8-bugbear
    #   flake8-builtins
    #   flake8-debugger
    #   flake8-print
flake8-black==0.3.6
    # via -r requirements.in
flake8-bugbear==24.12.12
    # via -r requirements.in
flake8-builtins==2.5.0
    # via -r requirements.in
flake8-debugger==4.1.2
    # via -r requirements.in
flake8-import-order==0.18.2
    # via -r requirements.in
flake8-print==5.0.0
    # via -r requirements.in
h11==0.14.0
    # via wsproto
identify==2.6.9
    # via pre-commit
idna==3.10
    # via
    #   requests
    #   trio
isort==5.10.1
    # via -r requirements.in
mccabe==0.7.0
    # via flake8
mypy-extensions==1.0.0
    # via black
nodeenv==1.9.1
    # via pre-commit
numpy==1.24.3
    # via
    #   -r requirements.in
    #   pandas
outcome==1.3.0.post0
    # via
    #   trio
    #   trio-websocket
packaging==24.2
    # via black
pandas==2.0.3
    # via -r requirements.in
pathspec==0.12.1
    # via black
platformdirs==4.3.6
    # via
    #   black
    #   virtualenv
pre-commit==4.0.1
    # via -r requirements.in
pycodestyle==2.12.1
    # via
    #   flake8
    #   flake8-debugger
    #   flake8-import-order
    #   flake8-print
pyflakes==3.2.0
    # via flake8
pysocks==1.7.1
    # via urllib3
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2025.1
    # via pandas
pyyaml==6.0.2
    # via pre-commit
requests==2.31.0
    # via -r requirements.in
selenium==4.11.0
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via trio
sortedcontainers==2.4.0
    # via trio
trio==0.29.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.12.2
    # via selenium
tzdata==2025.1
    # via pandas
urllib3==2.3.0
    # via
    #   requests
    #   selenium
virtualenv==20.29.3
    # via pre-commit
wsproto==1.2.0
    # via trio-websocket

# The following packages are considered to be unsafe in a requirements file:
# setuptools
